requirements.txt
asgiref==3.7.2
Django==3.2.9
django-taggit==4.0.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
